reorganize use statements, add documentation to submodules
